resolve conflict in testbench

+++ testbench.py
@@ -9,11 +9,12 @@ from env import TMP
 # from tests.cntr_parser_test import TestCntrParser
 # from tests.parser_integration_test import TestParserIntegration
 # from tests.mediautil_test import TestMediaUtil
-# from tests.mediautil_test import TestMediaUtil
 # from tests.mod_test import TestMod
+# from tests.sound_handler_test import TestSoundHandler
 
 
 if __name__ == "__main__":
+    if os.path.exists(TMP):
+        shutil.rmtree(TMP)
     os.mkdir(TMP)
     unittest.main()
-    shutil.rmtree(TMP)
